+++ packages/core/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [0.9.1](https://github.com/rayners/fvtt-seasons-and-stars/compare/seasons-and-stars-v0.9.0...seasons-and-stars-v0.9.1) (2025-07-23)
+
+
+### Bug Fixes
+
+* golarion calendar year offset bug for variants ([#154](https://github.com/rayners/fvtt-seasons-and-stars/issues/154)) ([80a9a81](https://github.com/rayners/fvtt-seasons-and-stars/commit/80a9a81c90e6d067299a7c33e7895463214217af))
+
+## [0.9.0](https://github.com/rayners/fvtt-seasons-and-stars/compare/seasons-and-stars-v0.8.0...seasons-and-stars-v0.9.0) (2025-07-21)
+
+
+### Features
+
+* add GM warning dialog for calendar deprecation ([#158](https://github.com/rayners/fvtt-seasons-and-stars/issues/158)) ([eda42dd](https://github.com/rayners/fvtt-seasons-and-stars/commit/eda42dde7ee2cf3affbd04e19d052af4bc09aa49))
+* add negative quick time button to default ([#152](https://github.com/rayners/fvtt-seasons-and-stars/issues/152)) ([5f866e5](https://github.com/rayners/fvtt-seasons-and-stars/commit/5f866e5d23042575aa4573f066e9629c920d3ef5))
+
+
+### Bug Fixes
+
+* eliminate ReDoS vulnerabilities by migrating to Handlebars helper architecture ([#155](https://github.com/rayners/fvtt-seasons-and-stars/issues/155)) ([54ce66b](https://github.com/rayners/fvtt-seasons-and-stars/commit/54ce66b3df0cc1cd585c0572704134da1ad0f5a4))
+
 ## [0.8.0](https://github.com/rayners/fvtt-seasons-and-stars/compare/seasons-and-stars-v0.7.0...seasons-and-stars-v0.8.0) (2025-07-18)